+++ pkg/azuremachine/mutate_create.go
@@ -4,71 +4,22 @@ import (
 	"context"
 
 	"github.com/giantswarm/microerror"
-	"github.com/giantswarm/micrologger"
-	"k8s.io/api/admission/v1beta1"
 	capz "sigs.k8s.io/cluster-api-provider-azure/api/v1alpha3"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/azure-admission-controller/internal/patches"
-	"github.com/giantswarm/azure-admission-controller/pkg/generic"
 	"github.com/giantswarm/azure-admission-controller/pkg/key"
 	"github.com/giantswarm/azure-admission-controller/pkg/mutator"
 )
 
-type CreateMutator struct {
-	ctrlClient client.Client
-	location   string
-	logger     micrologger.Logger
-}
-
-type CreateMutatorConfig struct {
-	CtrlClient client.Client
-	Location   string
-	Logger     micrologger.Logger
-}
-
-func NewCreateMutator(config CreateMutatorConfig) (*CreateMutator, error) {
-	if config.Logger == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
-	}
-	if config.CtrlClient == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.CtrlClient must not be empty", config)
-	}
-	if config.Location == "" {
-		return nil, microerror.Maskf(invalidConfigError, "%T.Location must not be empty", config)
-	}
-
-	m := &CreateMutator{
-		ctrlClient: config.CtrlClient,
-		location:   config.Location,
-		logger:     config.Logger,
-	}
-
-	return m, nil
-}
-
-func (m *CreateMutator) Mutate(ctx context.Context, request *v1beta1.AdmissionRequest) ([]mutator.PatchOperation, error) {
+func (h *WebhookHandler) OnCreateMutate(ctx context.Context, object interface{}) ([]mutator.PatchOperation, error) {
 	var result []mutator.PatchOperation
-
-	if request.DryRun != nil && *request.DryRun {
-		m.logger.LogCtx(ctx, "level", "debug", "message", "Dry run is not supported. Request processing stopped.")
-		return result, nil
-	}
-
-	azureMachineCR := &capz.AzureMachine{}
-	if _, _, err := mutator.Deserializer.Decode(request.Object.Raw, nil, azureMachineCR); err != nil {
-		return []mutator.PatchOperation{}, microerror.Maskf(parsingFailedError, "unable to parse AzureMachine CR: %v", err)
-	}
-
-	capi, err := generic.IsCAPIRelease(azureMachineCR)
+	azureMachineCR, err := key.ToAzureMachinePtr(object)
 	if err != nil {
 		return []mutator.PatchOperation{}, microerror.Mask(err)
 	}
-	if capi {
-		return []mutator.PatchOperation{}, nil
-	}
+	azureMachineCROriginal := azureMachineCR.DeepCopy()
 
-	patch, err := m.ensureLocation(ctx, azureMachineCR)
+	patch, err := h.ensureLocation(ctx, azureMachineCR)
 	if err != nil {
 		return []mutator.PatchOperation{}, microerror.Mask(err)
 	}
@@ -76,7 +27,7 @@ func (m *CreateMutator) Mutate(ctx context.Context, request *v1beta1.AdmissionRe
 		result = append(result, *patch)
 	}
 
-	patch, err = m.ensureOSDiskCachingType(ctx, azureMachineCR)
+	patch, err = h.ensureOSDiskCachingType(ctx, azureMachineCR)
 	if err != nil {
 		return []mutator.PatchOperation{}, microerror.Mask(err)
 	}
@@ -84,7 +35,7 @@ func (m *CreateMutator) Mutate(ctx context.Context, request *v1beta1.AdmissionRe
 		result = append(result, *patch)
 	}
 
-	patch, err = mutator.CopyAzureOperatorVersionLabelFromAzureClusterCR(ctx, m.ctrlClient, azureMachineCR.GetObjectMeta())
+	patch, err = mutator.CopyAzureOperatorVersionLabelFromAzureClusterCR(ctx, h.ctrlClient, azureMachineCR.GetObjectMeta())
 	if err != nil {
 		return []mutator.PatchOperation{}, microerror.Mask(err)
 	}
@@ -95,7 +46,7 @@ func (m *CreateMutator) Mutate(ctx context.Context, request *v1beta1.AdmissionRe
 	azureMachineCR.Default()
 	{
 		var capiPatches []mutator.PatchOperation
-		capiPatches, err = patches.GenerateFrom(request.Object.Raw, azureMachineCR)
+		capiPatches, err = patches.GenerateFromObjectDiff(azureMachineCROriginal, azureMachineCR)
 		if err != nil {
 			return []mutator.PatchOperation{}, microerror.Mask(err)
 		}
@@ -108,25 +59,9 @@ func (m *CreateMutator) Mutate(ctx context.Context, request *v1beta1.AdmissionRe
 	return result, nil
 }
 
-func (m *CreateMutator) Log(keyVals ...interface{}) {
-	m.logger.Log(keyVals...)
-}
-
-func (m *CreateMutator) Resource() string {
-	return "azuremachine"
-}
-
-func (m *CreateMutator) ensureLocation(ctx context.Context, azureMachine *capz.AzureMachine) (*mutator.PatchOperation, error) {
+func (h *WebhookHandler) ensureLocation(_ context.Context, azureMachine *capz.AzureMachine) (*mutator.PatchOperation, error) {
 	if azureMachine.Spec.Location == "" {
-		return mutator.PatchAdd("/spec/location", m.location), nil
-	}
-
-	return nil, nil
-}
-
-func (m *CreateMutator) ensureOSDiskCachingType(ctx context.Context, azureMachine *capz.AzureMachine) (*mutator.PatchOperation, error) {
-	if len(azureMachine.Spec.OSDisk.CachingType) < 1 {
-		return mutator.PatchAdd("/spec/osDisk/cachingType", key.OSDiskCachingType()), nil
+		return mutator.PatchAdd("/spec/location", h.location), nil
 	}
 
 	return nil, nil