+++ pkg/azuremachine/validate_create.go
@@ -4,67 +4,16 @@ import (
 	"context"
 
 	"github.com/giantswarm/microerror"
-	"github.com/giantswarm/micrologger"
-	"k8s.io/api/admission/v1beta1"
-	capz "sigs.k8s.io/cluster-api-provider-azure/api/v1alpha3"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/azure-admission-controller/internal/errors"
-	"github.com/giantswarm/azure-admission-controller/internal/vmcapabilities"
 	"github.com/giantswarm/azure-admission-controller/pkg/generic"
-	"github.com/giantswarm/azure-admission-controller/pkg/validator"
+	"github.com/giantswarm/azure-admission-controller/pkg/key"
 )
 
-type CreateValidator struct {
-	ctrlClient client.Client
-	location   string
-	logger     micrologger.Logger
-	vmcaps     *vmcapabilities.VMSKU
-}
-
-type CreateValidatorConfig struct {
-	CtrlClient client.Client
-	Location   string
-	Logger     micrologger.Logger
-	VMcaps     *vmcapabilities.VMSKU
-}
-
-func NewCreateValidator(config CreateValidatorConfig) (*CreateValidator, error) {
-	if config.CtrlClient == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.CtrlClient must not be empty", config)
-	}
-	if config.Logger == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
-	}
-	if config.Location == "" {
-		return nil, microerror.Maskf(invalidConfigError, "%T.Location must not be empty", config)
-	}
-	if config.VMcaps == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.VMcaps must not be empty", config)
-	}
-
-	v := &CreateValidator{
-		ctrlClient: config.CtrlClient,
-		location:   config.Location,
-		logger:     config.Logger,
-		vmcaps:     config.VMcaps,
-	}
-
-	return v, nil
-}
-
-func (a *CreateValidator) Validate(ctx context.Context, request *v1beta1.AdmissionRequest) error {
-	cr := &capz.AzureMachine{}
-	if _, _, err := validator.Deserializer.Decode(request.Object.Raw, nil, cr); err != nil {
-		return microerror.Maskf(errors.ParsingFailedError, "unable to parse AzureMachine CR: %v", err)
-	}
-
-	capi, err := generic.IsCAPIRelease(cr)
+func (h *WebhookHandler) OnCreateValidate(ctx context.Context, object interface{}) error {
+	cr, err := key.ToAzureMachinePtr(object)
 	if err != nil {
 		return microerror.Mask(err)
-	}
-	if capi {
-		return nil
 	}
 
 	err = cr.ValidateCreate()
@@ -73,7 +22,7 @@ func (a *CreateValidator) Validate(ctx context.Context, request *v1beta1.Admissi
 		return microerror.Mask(err)
 	}
 
-	err = generic.ValidateOrganizationLabelContainsExistingOrganization(ctx, a.ctrlClient, cr)
+	err = generic.ValidateOrganizationLabelContainsExistingOrganization(ctx, h.ctrlClient, cr)
 	if err != nil {
 		return microerror.Mask(err)
 	}
@@ -83,12 +32,12 @@ func (a *CreateValidator) Validate(ctx context.Context, request *v1beta1.Admissi
 		return microerror.Mask(err)
 	}
 
-	err = validateLocation(*cr, a.location)
+	err = validateLocation(*cr, h.location)
 	if err != nil {
 		return microerror.Mask(err)
 	}
 
-	supportedAZs, err := a.vmcaps.SupportedAZs(ctx, cr.Spec.Location, cr.Spec.VMSize)
+	supportedAZs, err := h.vmcaps.SupportedAZs(ctx, cr.Spec.Location, cr.Spec.VMSize)
 	if err != nil {
 		return microerror.Mask(err)
 	}
@@ -99,8 +48,4 @@ func (a *CreateValidator) Validate(ctx context.Context, request *v1beta1.Admissi
 	}
 
 	return nil
-}
-
-func (a *CreateValidator) Log(keyVals ...interface{}) {
-	a.logger.Log(keyVals...)
 }