+++ pkg/azuremachine/webhook_handler.go
@@ -1,0 +1,87 @@
+package azuremachine
+
+import (
+	"context"
+
+	"github.com/giantswarm/microerror"
+	"github.com/giantswarm/micrologger"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	capz "sigs.k8s.io/cluster-api-provider-azure/api/v1alpha3"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	"github.com/giantswarm/azure-admission-controller/internal/errors"
+	"github.com/giantswarm/azure-admission-controller/internal/vmcapabilities"
+	"github.com/giantswarm/azure-admission-controller/pkg/key"
+	"github.com/giantswarm/azure-admission-controller/pkg/mutator"
+	"github.com/giantswarm/azure-admission-controller/pkg/validator"
+)
+
+type WebhookHandler struct {
+	ctrlClient client.Client
+	decoder    runtime.Decoder
+	location   string
+	logger     micrologger.Logger
+	vmcaps     *vmcapabilities.VMSKU
+}
+
+type WebhookHandlerConfig struct {
+	CtrlClient client.Client
+	Decoder    runtime.Decoder
+	Location   string
+	Logger     micrologger.Logger
+	VMcaps     *vmcapabilities.VMSKU
+}
+
+func NewWebhookHandler(config WebhookHandlerConfig) (*WebhookHandler, error) {
+	if config.CtrlClient == nil {
+		return nil, microerror.Maskf(invalidConfigError, "%T.CtrlClient must not be empty", config)
+	}
+	if config.Decoder == nil {
+		return nil, microerror.Maskf(invalidConfigError, "%T.Decoder must not be empty", config)
+	}
+	if config.Logger == nil {
+		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
+	}
+	if config.Location == "" {
+		return nil, microerror.Maskf(invalidConfigError, "%T.Location must not be empty", config)
+	}
+	if config.VMcaps == nil {
+		return nil, microerror.Maskf(invalidConfigError, "%T.VMcaps must not be empty", config)
+	}
+
+	v := &WebhookHandler{
+		ctrlClient: config.CtrlClient,
+		decoder:    config.Decoder,
+		location:   config.Location,
+		logger:     config.Logger,
+		vmcaps:     config.VMcaps,
+	}
+
+	return v, nil
+}
+
+func (h *WebhookHandler) Log(keyVals ...interface{}) {
+	h.logger.Log(keyVals...)
+}
+
+func (h *WebhookHandler) Resource() string {
+	return "azuremachine"
+}
+
+func (h *WebhookHandler) Decode(rawObject runtime.RawExtension) (metav1.ObjectMetaAccessor, error) {
+	cr := &capz.AzureMachine{}
+	if _, _, err := validator.Deserializer.Decode(rawObject.Raw, nil, cr); err != nil {
+		return nil, microerror.Maskf(errors.ParsingFailedError, "unable to parse AzureMachine CR: %v", err)
+	}
+
+	return cr, nil
+}
+
+func (h *WebhookHandler) ensureOSDiskCachingType(_ context.Context, azureMachine *capz.AzureMachine) (*mutator.PatchOperation, error) {
+	if len(azureMachine.Spec.OSDisk.CachingType) < 1 {
+		return mutator.PatchAdd("/spec/osDisk/cachingType", key.OSDiskCachingType()), nil
+	}
+
+	return nil, nil
+}